==> 01_data_cleaning.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

#TODO 
 - Text kürzen
 - Reproduce Plots
 - Do a Map (Maps are alwas important!)
 - PCA beschreiben

```{r}
#pacman::p_install_gh("systats/binoculaR")
#devtools::install_github("strengejacke/sjlabelled",dependencies=TRUE)
#devtools::install_github("strengejacke/sjmisc",dependencies=TRUE)
#devtools::install_github("strengejacke/sjstats",dependencies=TRUE)
#devtools::install_github("strengejacke/ggeffects",dependencies=TRUE)
#devtools::install_github("strengejacke/sjPlot",dependencies=TRUE)
#devtools::install_github("lme4/lme4",dependencies=TRUE)
#pacman::p_install_gh("systats/binoculaR")

pacman::p_load(tidyverse, haven, magrittr, car, psych, sjPlot, sjstats, sjmisc, lme4, binoculaR, MuMIn)


```

# Load in Data

```{r}
arab4 <- read_spss("data/arab4.sav")
arab3 <- read_spss("data/arab3.sav")

# arab4 <- get(load(url("https://github.com/favstats/GodlyGovernance/raw/master/data/arab4.Rdata")))


```

# Inspect Data

```{r}
#b_select <- binoculaR::binoculaR(arab4)
#dput(b_select$var_codes)

#index <- c("country","wt" ,"sample" ,"a1", "q1", "q2", "q13", "q5182", 
#"q5184", "q6013", "q6014", "q6018", "q60118", "q605a", "q6061", "q6062", 
#"q6063", "q6064", "q6071", "q6074", "q6076", "q707", "q708", 
#"q7114", "q1001", "q1002", "q1003", "t1003", "q1004", "q1005", 
#"q609", "q6101", "q6106", "q1012", "q1012a", "q1016")
#
#arab4 %<>%
#  select(index)
#
#arab4
#view_df2(arab3, hide.progress = T)
# save(arab4, file = "data/arab4.Rdata")
```

# Filter Data

```{r}
arab4 %<>%
  filter(q1012 == 1) %>%#only Muslims
  mutate(sample = ifelse(is.na(sample) | sample == 1, 1, 2)) %>% 
  filter(sample != 2) #only non-refugees

arab3 %<>%
  filter(q1012 == 1) #only Muslims

arab3
arab4



```

# Recode Data

## Arab3

```{r}

arab3 %<>% 
  mutate(cntry = sjmisc::to_label(country)) %>% 
  mutate(region = sjmisc::to_label(a1)) %>% 
  mutate(governorate = sjmisc::to_label(q1)) %>% 
  mutate(year = lubridate::year(date))

  # Dependent Variable
arab3 %<>% 
  mutate(islamistparties1 = ifelse(q518a2 >= 5, NA, 5 - q518a2)) %>%
  mutate(islamistparties2 = ifelse(q518b2 >= 5, NA, 5 - q518b2)) %>% 
  mutate(islamistparties = case_when(
        islamistparties1 == 1 ~ 1,
        islamistparties1 == 2 ~ 2,
        islamistparties1 == 3 ~ 3,
        islamistparties1 == 4 ~ 4,
        islamistparties2 == 1 ~ 1,
        islamistparties2 == 2 ~ 2,
        islamistparties2 == 3 ~ 3,
        islamistparties2 == 4 ~ 4)
    ) %>%  
  mutate(islamistgov = ifelse(q5184 >= 5, NA, 5 - q5184)) %>% 
  mutate(religinterfere = ifelse(q6061 > 5, NA, q6061)) %>% 
  mutate(religleaders = ifelse(q6062 >= 5, NA, 5 - q6062)) %>% 
  mutate(religleadersinfl = ifelse(q6063 >= 5, NA, 5 - q6063)) %>%
  mutate(seperation = ifelse(q6064 > 5, NA, q6064)) %>% 
  mutate(religparty = Recode(q605a, "1 = 1;
                                     2 = 1;
                                     3 = 0;
                                     4 = 0;
                                     5 = 0;
                                    8 = 0;
                                    9 = NA")) %>% 
  mutate(religparty2 = Recode(q605a, "1 = 5;
                                      2 = 4;
                                      3 = 2;
                                      4 = 1;
                                      5 = 3;
                                     8 = NA;
                                     9 = NA"))

ifelse4cat_rec <- function(variable) {
  recoded <- ifelse(variable == 0 | variable > 5, NA, 5 - variable)
  return(recoded)
}


arab3 %<>% 
  mutate(female = ifelse(sex == 2, 1, 0)) %>% 
  mutate(work = ifelse(q1004 == 0 | q1004 > 5, NA, abs(q1004 - 2))) %>% 
  mutate(income = ifelse4cat_rec(q1016)) %>% 
  mutate(age = ifelse(q1001 == 0 | q1001 == 9999 | q1001 == 999, NA, q1001)) %>% 
  mutate(educ = case_when(
        q1003 == 0 ~ NA_real_,
        q1003 == 99 ~ NA_real_,
        q1003 == 5 ~ 4,
        q1003 == 6 ~ 5,
        q1003 == 7 ~ 6,
        q1003yem == 0 ~ NA_real_,
        q1003yem == 99 ~ NA_real_,
        q1003yem == 4 ~ 3,
        q1003yem == 5 ~ 4,
        q1003yem == 6 ~ 5,
        q1003yem == 7 ~ 5,
        q1003yem == 8 ~ 6,
        q1003t == 0 ~ NA_real_,
        q1003t == 99 ~ NA_real_,
        q1003t == 1 ~ 1,
        q1003t == 2 ~ 2,
        q1003t == 3 ~ 3,
        q1003t == 4 ~ 4,
        q1003t == 5 ~ 5,
        q1003t == 6 ~ 6,
    TRUE ~ as.numeric(q1003))
    ) %>% 
  mutate(inf_us = ifelse(q7011 == 0 | q7011 > 5, NA, q7011)) %>% 
  mutate(inf_eu = ifelse(q7012 == 0 | q7012 > 5, NA, q7012)) %>% 
  mutate(globalism = ifelse(q701b == 0 | q701b > 5, NA, q701b)) %>% 
  mutate(pray = ifelse(q6101 == 0 | q6101 > 5, NA, 6 - q6101)) %>% 
  mutate(quran = ifelse(q6106 == 0 | q6106 > 5, NA, 6 - q6106)) %>%
  mutate(womanwork = ifelse(q6012 == 0 | q6012 > 5, NA, q6012)) %>% 
  mutate(womenleader = ifelse4cat_rec(q6013)) %>% 
  mutate(womeneduc = ifelse4cat_rec(q6014)) %>% 
  mutate(nodemoc = ifelse(q6071 == 0 | q6071 > 5, NA, q6071)) %>% 
  mutate(genderapartuni = ifelse4cat_rec(q6074)) %>% 
  mutate(coverup = ifelse4cat_rec(q6076)) %>% 
  select(cntry, year, region, governorate, islamistparties , islamistgov, religinterfere, religleaders, religleadersinfl, seperation, religparty, religparty2, female, work, income,  age, educ, globalism, pray, quran, womanwork, womenleader, womeneduc, nodemoc, genderapartuni, coverup, inf_us, inf_eu, wt)





```

## Arab4

```{r}
table(arab4$country)

arab4 %<>% 
  mutate(cntry = sjmisc::to_label(country)) %>% 
  mutate(region = sjmisc::to_label(a1)) %>% 
  mutate(governorate = sjmisc::to_label(q1)) %>% 
  mutate(year = 2016)#%>%
#  mutate(district = sjmisc::to_label(q2)) 

# Dependent Variable
arab4 %<>% 
  mutate(islamistparties = ifelse(q5182 >= 5, NA, 5 - q5182)) %>% 
  mutate(islamistgov = ifelse(q5184 >= 5, NA, 5 - q5184)) %>% 
  mutate(religinterfere = ifelse(q6061 > 5, NA, q6061)) %>% 
  mutate(religleaders = ifelse(q6062 >= 5, NA, 5 - q6062)) %>% 
  mutate(religleadersinfl = ifelse(q6063 >= 5, NA, 5 - q6063)) %>%
  mutate(seperation = ifelse(q6064 > 5, NA, q6064)) %>% 
  mutate(religparty = Recode(q605a, "1 = 1;
                                     2 = 1;
                                     3 = 0;
                                     4 = 0;
                                     5 = 0;
                                    98 = 0;
                                    99 = NA")) %>% 
  mutate(religparty2 = Recode(q605a, "1 = 5;
                                      2 = 4;
                                      3 = 2;
                                      4 = 1;
                                      5 = 3;
                                     98 = NA;
                                     99 = NA"))



arab4 %<>% 
  mutate(female = ifelse(q1002 == 2, 1, 0)) %>% 
  mutate(work = ifelse(q1004 == 0 | q1004 > 5, NA, abs(q1004 - 2))) %>% 
  mutate(income = ifelse4cat_rec(q1016)) %>% 
  mutate(age = ifelse(q1001 == 0 | q1001 == 9999 | q1001 == 999, NA, q1001)) %>% 
  mutate(educ = case_when(
        q1003 == 0 ~ NA_real_,
        q1003 == 98 ~ NA_real_,
        q1003 == 99 ~ NA_real_,
        q1003 == 5 ~ 4,
        q1003 == 6 ~ 5,
        q1003 == 7 ~ 6,
        t1003 == 0 ~ NA_real_,
        t1003 == 98 ~ NA_real_,
        t1003 == 99 ~ NA_real_,
        t1003 == 3 ~ 3,
        t1003 == 4 ~ 4,
        t1003 == 5 ~ 5,
        t1003 == 6 ~ 6,
    TRUE ~ as.numeric(q1003))
    ) %>% 
  mutate(inf_us = ifelse(q7011 == 0 | q7011 > 5, NA, q7011)) %>% 
  mutate(inf_eu = ifelse(q7012 == 0 | q7012 > 5, NA, q7012)) %>% 
  mutate(globalism = ifelse(q701b == 0 | q701b > 5, NA, q701b)) %>% 
  mutate(pray = ifelse(q6101 == 0 | q6101 > 5, NA, 6 - q6101)) %>% 
  mutate(quran = ifelse(q6106 == 0 | q6106 > 5, NA, 6 - q6106)) %>% 
  mutate(womanwork = ifelse(q6012 == 0 | q6012 > 5, NA, q6012)) %>% 
  mutate(womenleader = ifelse4cat_rec(q6013)) %>% 
  mutate(womeneduc = ifelse4cat_rec(q6014)) %>% 
  mutate(nodemoc = ifelse(q6071 == 0 | q6071 > 5, NA, q6071)) %>% 
  mutate(genderapartuni = ifelse4cat_rec(q6074)) %>% 
  mutate(coverup = ifelse4cat_rec(q6076)) %>% 
  select(cntry, year, region, governorate , islamistparties , islamistgov, religinterfere, religleaders, religleadersinfl, seperation, religparty, religparty2, female, work, income,  age, educ, globalism, pray, quran, womanwork, womenleader, womeneduc, nodemoc, genderapartuni, coverup, inf_us, inf_eu, wt)

```


# Merging

```{r}
arab3;arab4

arab <- rbind(arab4,arab3) %>% 
  filter(4 >= islamistgov |4 >= religleaders)

arab

table(arab$age)
table(arab$cntry)
```



# Factor Analysis

```{r}
# f1 <- arab %>% 
#   select(islamistparties, islamistgov, 
#          religleaders, religleadersinfl) %>%
#   # na.omit() %>% 
#   psych::fa(1, rotate = "varimax",   
#         fm = "pa",
#         scores = "regression",
#         weight = na.omit(arab$wt)  )         
# fa.diagram(f1)                                  
# f1              
load("data/arab_reg.Rdata")
library(tidyverse)
arab_reg %>% 
  select(islamistparties, islamistgov, 
         religleaders, religleadersinfl) %>% 
  psych::pca(weight = arab_reg$wt, rotate = "oblimin") -> f2

data.frame(loadings = f2$loadings[1:4]) %>% 
  mutate(loadings = round(loadings, 2)) %>% 
  mutate(variable = c("Only Islamist Parties",
                      "Islamist Government (no elections)",
                      "Religious Leaders hold office",
                      "Religious Leaders influence decisions")) %>% 
  ggplot(aes(variable, y = 1, fill = loadings)) +
    geom_tile() +
  coord_flip() +
  ggthemes::theme_hc() +
  scale_fill_continuous(low = "lightblue", high = "blue") +
   geom_text(aes(variable, y = 1, label = loadings), color = "black", size = 4)  +
   geom_text(aes(4.3, y = 1.35, 
                 label = "Cronbach~~alpha==0.70"), 
             color = "black", size = 5, parse = T)   +
   geom_text(aes(3.7, y = 1.35, 
                 label = "KMO = 0.68"), 
             color = "black", size = 5)    +
   geom_text(aes(0.6, y = 1.35, 
                 label = "Variance explained = 53%"), 
             color = "black", size = 5) +
  xlab("") +
  ylab("") +
  guides(fill = F) +
  theme(axis.ticks.x = element_blank(), axis.text.x = element_blank())

ggsave(filename = "images/factor.png", width = 8, height = 4)
  
  

arab %>% 
  select(islamistparties, islamistgov, 
         religleaders, religleadersinfl) %>% 
  alpha() 

arab %>% 
  select(islamistparties, islamistgov, 
         religleaders, religleadersinfl) %>% 
  KMO() 


arab4 %>% 
  select(islamistparties, islamistgov, 
         religleaders, religleadersinfl,
         religparty2, seperation,
         religinterfere) %>% 
  na.omit() %>% 
  cor()  

# arab <- predict.psych(f1, arab %>% 
#   select(islamistparties, islamistgov, 
#          religleaders, religleadersinfl)) %>% 
#   tbl_df() %>% 
#   cbind(arab) %>% 
#   mutate(islamism_fa = PA1)

arab <- predict.psych(f2, arab %>% 
  select(islamistparties, islamistgov, 
         religleaders, religleadersinfl)) %>% 
  tbl_df() %>% 
  transmute(islamism = PC1) %>% 
  cbind(arab) 

#%>% 
#  ggplot(aes(islamism)) +
#  geom_histogram() +
#  facet_wrap(~cntry, scales = "free")

#sjPlot::view_df(arab, show.frq = T, show.prc = T)
# 
# cor.test(arab$PC1, arab$PA1)

save(arab, file = "data/arab.Rdata")
load("data/arab.Rdata") 
nrow(arab)
```

# Creating indices

```{r}
f1 <- arab %>% 
  select(womanwork, womenleader, womeneduc) %>% 
  fa(1, rotate = "promax",   
        fm = "pa",
        scores = "regression")           
fa.diagram(f1)                                  
f1        

arab %>% 
  select(womanwork, womenleader, womeneduc) %>% 
  psych::pca() 

f2 <- arab %>% 
  select(pray, quran) %>% 
  fa(1, rotate = "promax",   
        fm = "pa",
        scores = "regression")           
fa.diagram(f2)                                  
f2    

f3 <- arab %>% 
  select(nodemoc, genderapartuni, coverup) %>% 
  fa(1, rotate = "promax",   
        fm = "pa",
        scores = "regression")           
fa.diagram(f3)                                  
f3    

arab %>% 
  select(nodemoc, genderapartuni, coverup) %>% 
  KMO()

arab %>% 
  select(nodemoc, genderapartuni, coverup) %>% 
  psych::pca()

arab %>% 
  select(inf_eu, inf_us) %>% 
  psych::pca()


range01 <- function(x){(x - min(x, na.rm = T)) / (max(x, na.rm = T) - min(x, na.rm = T))}


arab_reg <- arab %>% 
  mutate(patriarchalvalues = range01(womanwork + womenleader + womeneduc))  %>% 
  mutate(personalpiety = range01(pray + quran)) %>% 
  mutate(antiwestern = range01(inf_eu + inf_us)) %>% 
  mutate(islamism = range01(islamism)) %>% 
  mutate_at(vars(female:inf_eu), range01) %>% 
#  mutate(islamism = scale(islamism)) %>% 
#  mutate(islamism = scale(islamism)) %>% 
#  mutate_if(.predicate = is.double, scale) 
  mutate(liberalislam = range01(nodemoc + genderapartuni + coverup)) %>% 
  mutate(cntryears = paste(cntry, year)) %>% 
  mutate(year_2016 = ifelse(year == 2016, 1, 0)) %>% 
  mutate(year_2014 = ifelse(year == 2014, 1, 0)) %>% 
  mutate(year_2013 = ifelse(year == 2013, 1, 0)) %>% 
  mutate(year_2012 = ifelse(year == 2012, 1, 0)) %>% 
  select(cntry, islamism, female, work, income, age, educ, 
                 year_2012, year_2013, year_2014, 
         islamistparties:islamistgov,
         religleaders:religleadersinfl,
                 globalism, personalpiety, patriarchalvalues, liberalislam,
                 wt)  %>% 
  filter(4 >= islamistgov) %>%
  filter(4 >= religleaders)%>% 
  na.omit()

# arab_reg$islamism_fa <- as.numeric(range01(arab_reg$islamism_fa) * 100)
# arab_reg$islamism_pca <- as.numeric(range01(arab_reg$islamism_pca) * 100)
# 
# arab_reg$income             <- as.numeric(scale(arab_reg$income))
# arab_reg$age                <- as.numeric(scale(arab_reg$age))
# arab_reg$educ               <- as.numeric(scale(arab_reg$educ)) 
# arab_reg$globalism          <- as.numeric(scale(arab_reg$globalism))   
# arab_reg$personalpiety      <- as.numeric(scale(arab_reg$personalpiety))       
# arab_reg$patriarchalvalues  <- as.numeric(scale(arab_reg$patriarchalvalues))           
# arab_reg$liberalislam       <- as.numeric(scale(arab_reg$liberalislam))     
# 
# hist(arab_reg$liberalislam)
# 
# 
# arab_reg %>% 
#   select(islamism_pca , female , work , income , age , educ , globalism , personalpiety , patriarchalvalues , liberalislam, cntry) %>% 
#   descr()
#   na.omit() %>% 
#   select(cntry) %>% 
#   table()

table(arab$year, arab$cntry)

save(arab_reg, file = "data/arab_reg.Rdata")

table(arab_reg$cntry)
```


# Regression

```{r}
model0 <- arab_reg %>% 
  lme4::lmer(islamism ~ 1 + (1|cntry), data = .)

icc(model0)

model1a <- lme4::lmer(islamism_pca ~ female + work + income + age + educ + globalism + personalpiety + patriarchalvalues + nodemoc + genderapartuni + coverup + (1|cntry), data = arab_reg)

model1 <- lme4::lmer(islamism ~ female + work + income + age + educ + globalism + antiwestern + personalpiety + patriarchalvalues + liberalislam + year_2012 + year_2013 + year_2014 + (1|cntry), data = arab_reg, weights = wt)

texreg::screenreg(model1)

anova(model1a, model1b)

r.squaredGLMM(model1a)

plot_model(model1, type = "re", sort.est = T, show.values = T, show.p = T, value.offset = 0.5)

#plot_model(model1, type = "std", sort.est = T, show.values = T, show.p = T, value.offset = 0.5)

sjp.lmer(model1, type = "fe.std", sort.est = T, show.values = T, show.p = T, y.offset = 0.4, p.kr = FALSE)

plot_model(model1, terms = c("personalpiety"), type = "eff")

icc(model1)

model2 <- lme4::lmer(islamism_pca ~ female + work + income + age + educ + globalism + personalpiety + patriarchalvalues + liberalislam + liberalislam*personalpiety + (1|cntry/year), data = arab_reg)

texreg::screenreg(model2)


plot_model(model2, sort.est = T, show.values = T, show.p = T, value.offset = 0.5)

sjp.int(model2, p.kr = FALSE, mdrt.values = "all")

sjp.lmer(model2, type = "fe.std", sort.est = T, show.values = T, show.p = T, y.offset = 0.4, p.kr = FALSE)


#plot_model(model2, type = "int")

vif.mer <- function (fit) {
  ## adapted from rms::vif
  
  v <- vcov(fit)
  nam <- names(fixef(fit))
  
  ## exclude intercepts
  ns <- sum(1 * (nam == "Intercept" | nam == "(Intercept)"))
  if (ns > 0) {
    v <- v[-(1:ns), -(1:ns), drop = FALSE]
    nam <- nam[-(1:ns)]
  }
  
  d <- diag(v)^0.5
  v <- diag(solve(v/(d %o% d)))
  names(v) <- nam
  v
}

vif.mer(model2)



model3 <- lme4::lmer(islamism_pca ~ female + work + income + age + educ + globalism + personalpiety + patriarchalvalues + nodemoc + genderapartuni + coverup + patriarchalvalues*personalpiety + (1|cntry/year), data = arab_reg)

texreg::screenreg(model3)


plot_model(model3, sort.est = T, show.values = T, show.p = T, value.offset = 0.5)

sjp.int(model3, p.kr = FALSE, mdrt.values = "all")

sjp.lmer(model3, type = "fe.std", sort.est = T, show.values = T, show.p = T, y.offset = 0.4, p.kr = FALSE)

```


```{r}
model1 <- arab_reg %>% 
  lme4::lmer(islamism ~ female + work + income + age + educ + globalism + pray + quran + womanwork + womenleader + womeneduc + nodemoc + genderapartuni + coverup + (1|cntry/year), data = .)

texreg::screenreg(model1)

plot_model(model1, sort.est = T, show.values = T, show.p = T, value.offset = 0.5)


plot_model(model1, terms = c("nodemoc", "female"), type = "pred")

model2 <- arab_reg %>% 
  lme4::lmer(islamism ~ female + work + income + age + educ + globalism + pray + quran + womanwork + womenleader + womeneduc + nodemoc + genderapartuni + coverup + nodemoc*quran + (1|cntry/year), data = .)

texreg::screenreg(model2)
```


```{r}
view_df2 <- function (x, weight.by = NULL, altr.row.col = TRUE, show.id = TRUE, 
  show.type = FALSE, show.values = TRUE, show.string.values = FALSE, 
  show.labels = TRUE, show.frq = FALSE, show.prc = FALSE, 
  show.wtd.frq = FALSE, show.wtd.prc = FALSE, show.na = FALSE, 
  max.len = 15, sort.by.name = FALSE, wrap.labels = 50, hide.progress = FALSE, 
  CSS = NULL, encoding = NULL, file = NULL, use.viewer = TRUE, 
  no.output = FALSE, remove.spaces = TRUE) 
{
  get.encoding <- function(encoding, data = NULL) {
  if (is.null(encoding)) {
    if (!is.null(data) && is.data.frame(data)) {
      # get variable label
      labs <- sjlabelled::get_label(data[[1]])
      # check if vectors of data frame have
      # any valid label. else, default to utf-8
      if (!is.null(labs) && is.character(labs))
        encoding <- Encoding(sjlabelled::get_label(data[[1]]))
      else
        encoding <- "UTF-8"
      # unknown encoding? default to utf-8
      if (encoding == "unknown") encoding <- "UTF-8"
    } else if (.Platform$OS.type == "unix")
      encoding <- "UTF-8"
    else
      encoding <- "Windows-1252"
  }
  return(encoding)
}

  has_value_labels <- function(x) {
  !(is.null(attr(x, "labels", exact = T)) && is.null(attr(x, "value.labels", exact = T)))
}

  sju.rmspc <- function(html.table) {
  cleaned <- gsub("      <", "<", html.table, fixed = TRUE, useBytes = TRUE)
  cleaned <- gsub("    <", "<", cleaned, fixed = TRUE, useBytes = TRUE)
  cleaned <- gsub("  <", "<", cleaned, fixed = TRUE, useBytes = TRUE)
  return(cleaned)
  }
  
  frq.value <- function(index, x, df.val, weights = NULL) {
  valstring <- ""
  # check if we have a valid index
  if (index <= ncol(x) && !is.null(df.val[[index]])) {
    # do we have weights?
    if (!is.null(weights))
      variab <- sjstats::weight(x[[index]], weights)
    else
      variab <- x[[index]]
    # create frequency table. same function as for
    # sjt.frq and sjp.frq
    ftab <- create.frq.df(variab, 20)$mydat$frq
    # remove last value, which is N for NA
    if (length(ftab) == 1 && is.na(ftab)) {
      valstring <- "<NA>"
    } else {
      for (i in 1:(length(ftab) - 1)) {
        valstring <- paste0(valstring, ftab[i])
        if (i < length(ftab)) valstring <- paste0(valstring, "<br>")
      }
    }
  } else {
    valstring <- ""
  }
  return(valstring)
}

prc.value <- function(index, x, df.val, weights = NULL) {
  valstring <- ""
  # check for valid indices
  if (index <= ncol(x) && !is.null(df.val[[index]])) {
    # do we have weights?
    if (!is.null(weights))
      variab <- sjstats::weight(x[[index]], weights)
    else
      variab <- x[[index]]
    # create frequency table, but only get valid percentages
    ftab <- create.frq.df(variab, 20)$mydat$valid.prc
    # remove last value, which is a NA dummy
    if (length(ftab) == 1 && is.na(ftab)) {
      valstring <- "<NA>"
    } else {
      for (i in 1:(length(ftab) - 1)) {
        valstring <- paste0(valstring, sprintf("%.2f", ftab[i]))
        if (i < length(ftab)) valstring <- paste0(valstring, "<br>")
      }
    }
  } else {
    valstring <- ""
  }
  return(valstring)
}
  
  encoding <- get.encoding(encoding, x)
  if (!is.data.frame(x)) 
    stop("Parameter needs to be a data frame!", call. = FALSE)
  df.var <- sjlabelled::get_label(x)
  df.val <- sjlabelled::get_labels(x)
  colcnt <- ncol(x)
  id <- seq_len(colcnt)
  if (sort.by.name) 
    id <- id[order(colnames(x))]
  tag.table <- "table"
  tag.thead <- "thead"
  tag.tdata <- "tdata"
  tag.arc <- "arc"
  tag.caption <- "caption"
  tag.omit <- "omit"
  css.table <- "border-collapse:collapse; border:none;"
  css.thead <- "border-bottom:double; font-style:italic; font-weight:normal; padding:0.2cm; text-align:left; vertical-align:top;"
  css.tdata <- "padding:0.2cm; text-align:left; vertical-align:top;"
  css.arc <- "background-color:#eeeeee"
  css.caption <- "font-weight: bold; text-align:left;"
  css.omit <- "color:#999999;"
  if (!is.null(CSS)) {
    if (!is.null(CSS[["css.table"]])) 
      css.table <- ifelse(substring(CSS[["css.table"]], 
        1, 1) == "+", paste0(css.table, substring(CSS[["css.table"]], 
        2)), CSS[["css.table"]])
    if (!is.null(CSS[["css.thead"]])) 
      css.thead <- ifelse(substring(CSS[["css.thead"]], 
        1, 1) == "+", paste0(css.thead, substring(CSS[["css.thead"]], 
        2)), CSS[["css.thead"]])
    if (!is.null(CSS[["css.tdata"]])) 
      css.tdata <- ifelse(substring(CSS[["css.tdata"]], 
        1, 1) == "+", paste0(css.tdata, substring(CSS[["css.tdata"]], 
        2)), CSS[["css.tdata"]])
    if (!is.null(CSS[["css.arc"]])) 
      css.arc <- ifelse(substring(CSS[["css.arc"]], 1, 
        1) == "+", paste0(css.arc, substring(CSS[["css.arc"]], 
        2)), CSS[["css.arc"]])
    if (!is.null(CSS[["css.caption"]])) 
      css.caption <- ifelse(substring(CSS[["css.caption"]], 
        1, 1) == "+", paste0(css.caption, substring(CSS[["css.caption"]], 
        2)), CSS[["css.caption"]])
    if (!is.null(CSS[["css.omit"]])) 
      css.omit <- ifelse(substring(CSS[["css.omit"]], 
        1, 1) == "+", paste0(css.omit, substring(CSS[["css.omit"]], 
        2)), CSS[["css.omit"]])
  }
  page.style <- sprintf("<style>\nhtml, body { background-color: white; }\n%s { %s }\n.%s { %s }\n.%s { %s }\n.%s { %s }\n%s { %s }\n.%s { %s }\n</style>", 
    tag.table, css.table, tag.thead, css.thead, tag.tdata, 
    css.tdata, tag.arc, css.arc, tag.caption, css.caption, 
    tag.omit, css.omit)
  toWrite <- sprintf("<html>\n<head>\n<meta http-equiv=\"Content-type\" content=\"text/html;charset=%s\">\n%s\n</head>\n<body>\n", 
    encoding, page.style)
  page.content <- sprintf("<table>\n  <caption>Data frame: %s</caption>\n", 
    deparse(substitute(x)))
  page.content <- paste0(page.content, "  <tr>\n    ")
  if (show.id) 
    page.content <- paste0(page.content, "<th class=\"thead\">ID</th>")
  page.content <- paste0(page.content, "<th class=\"thead\">Name</th>")
  if (show.type) 
    page.content <- paste0(page.content, "<th class=\"thead\">Type</th>")
  page.content <- paste0(page.content, "<th class=\"thead\">Label</th>")
  if (show.na) 
    page.content <- paste0(page.content, "<th class=\"thead\">missings</th>")
  if (show.values) 
    page.content <- paste0(page.content, "<th class=\"thead\">Values</th>")
  if (show.labels) 
    page.content <- paste0(page.content, "<th class=\"thead\">Value Labels</th>")
  if (show.frq) 
    page.content <- paste0(page.content, "<th class=\"thead\">Freq.</th>")
  if (show.prc) 
    page.content <- paste0(page.content, "<th class=\"thead\">%</th>")
  if (show.wtd.frq) 
    page.content <- paste0(page.content, "<th class=\"thead\">weighted Freq.</th>")
  if (show.wtd.prc) 
    page.content <- paste0(page.content, "<th class=\"thead\">weighted %</th>")
  page.content <- paste0(page.content, "\n  </tr>\n")
  if (!hide.progress) 
    pb <- utils::txtProgressBar(min = 0, max = colcnt, style = 3)
  for (ccnt in seq_len(colcnt)) {
    index <- id[ccnt]
    arcstring <- ""
    if (altr.row.col) 
      arcstring <- ifelse(sjmisc::is_even(ccnt), " arc", 
        "")
    page.content <- paste0(page.content, "  <tr>\n")
    if (show.id) 
      page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\">%i</td>\n", 
        arcstring, index))
    if (!is.list(x[[index]]) && !is.null(sjlabelled::get_note(x[[index]]))) 
      td.title.tag <- sprintf(" title=\"%s\"", sjlabelled::get_note(x[[index]]))
    else td.title.tag <- ""
    page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\"%s>%s</td>\n", 
      arcstring, td.title.tag, colnames(x)[index]))
    if (show.type) {
      vartype <- sjmisc::var_type(x[[index]])
      page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\">%s</td>\n", 
        arcstring, vartype))
    }
    if (index <= length(df.var)) {
      varlab <- df.var[index]
      if (!is.null(wrap.labels)) {
        varlab <- sjmisc::word_wrap(varlab, wrap.labels, 
          "<br>")
      }
    }
    else {
      varlab <- "<NA>"
    }
    page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\">%s</td>\n", 
      arcstring, varlab))
    if (show.na) {
      if (is.list(x[[index]])) {
        page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\"><span class=\"omit\">&lt;list&gt;</span></td>\n", 
          arcstring))
      }
      else {
        page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\">%i (%.2f%%)</td>\n", 
          arcstring, sum(is.na(x[[index]]), na.rm = T), 
          100 * sum(is.na(x[[index]]), na.rm = T)/nrow(x)))
      }
    }
    if (is.numeric(x[[index]]) && !has_value_labels(x[[index]])) {
      if (show.values || show.labels) {
        valstring <- paste0(sprintf("%a", range(x[[index]], 
          na.rm = T)), collapse = "-")
        if (show.values && show.labels) {
          colsp <- " colspan=\"2\""
          valstring <- paste0("<em>range: ", valstring, 
            "</em>")
        }
        else {
          colsp <- ""
        }
        page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\"%s>%s</td>\n", 
          arcstring, colsp, valstring))
      }
    }
    else {
      if (show.values) {
        valstring <- ""
        if (index <= ncol(x)) {
          if (is.list(x[[index]])) {
            valstring <- "<span class=\"omit\">&lt;list&gt;</span>"
          }
          else {
            vals <- sjlabelled::get_values(x[[index]])
            if (!is.null(vals)) {
              loop <- na.omit(seq_len(length(vals))[1:max.len])
              for (i in loop) {
                valstring <- paste0(valstring, vals[i])
                if (i < length(vals)) 
                  valstring <- paste0(valstring, "<br>")
              }
              if (max.len < length(vals)) 
                valstring <- paste0(valstring, "<span class=\"omit\">&lt;...&gt;</span>")
            }
          }
        }
        else {
          valstring <- "<NA>"
        }
        page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\">%s</td>\n", 
          arcstring, valstring))
      }
      if (show.labels) {
        valstring <- ""
        if (index <= length(df.val)) {
          if (is.list(x[[index]])) {
            valstring <- "<span class=\"omit\">&lt;list&gt;</span>"
          }
          else {
            vals <- df.val[[index]]
            if (!is.null(vals)) 
              vals <- na.omit(vals)
            if (is.character(x[[index]]) && !is.null(vals) && 
              !sjmisc::is_empty(vals)) {
              if (show.string.values) 
                vals <- sort(vals)
              else vals <- "<span class=\"omit\" title =\"'show.string.values = TRUE' to show values.\">&lt;output omitted&gt;</span>"
            }
            if (!is.null(vals)) {
              loop <- na.omit(seq_len(length(vals))[1:max.len])
              for (i in loop) {
                valstring <- paste0(valstring, vals[i])
                if (i < length(vals)) 
                  valstring <- paste0(valstring, "<br>")
              }
              if (max.len < length(vals)) 
                valstring <- paste0(valstring, "<span class=\"omit\">&lt;... truncated&gt;</span>")
            }
          }
        }
        else {
          valstring <- "<NA>"
        }
        page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\">%s</td>\n", 
          arcstring, valstring))
      }
    }
    if (show.frq) {
      if (is.list(x[[index]])) 
        valstring <- "<span class=\"omit\">&lt;list&gt;</span>"
      else valstring <- frq.value(index, x, df.val)
      page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\">%s</td>\n", 
        arcstring, valstring))
    }
    if (show.prc) {
      if (is.list(x[[index]])) 
        valstring <- "<span class=\"omit\">&lt;list&gt;</span>"
      else valstring <- prc.value(index, x, df.val)
      page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\">%s</td>\n", 
        arcstring, valstring))
    }
    if (show.wtd.frq && !is.null(weight.by)) {
      if (is.list(x[[index]])) 
        valstring <- "<span class=\"omit\">&lt;list&gt;</span>"
      else valstring <- frq.value(index, x, df.val, weight.by)
      page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\">%s</td>\n", 
        arcstring, valstring))
    }
    if (show.prc && !is.null(weight.by)) {
      if (is.list(x[[index]])) 
        valstring <- "<span class=\"omit\">&lt;list&gt;</span>"
      else valstring <- prc.value(index, x, df.val, weight.by)
      page.content <- paste0(page.content, sprintf("    <td class=\"tdata%s\">%s</td>\n", 
        arcstring, valstring))
    }
    if (!hide.progress) 
      utils::setTxtProgressBar(pb, ccnt)
    page.content <- paste0(page.content, "  </tr>\n")
  }
  if (!hide.progress) 
    close(pb)
  page.content <- paste(page.content, "</table>", sep = "\n")
  toWrite <- paste0(toWrite, sprintf("%s\n</body></html>", 
    page.content))
  knitr <- page.content
  knitr <- gsub("class=", "style=", knitr, fixed = TRUE, useBytes = TRUE)
  knitr <- gsub("<table", sprintf("<table style=\"%s\"", css.table), 
    knitr, fixed = TRUE, useBytes = TRUE)
  knitr <- gsub(tag.tdata, css.tdata, knitr, fixed = TRUE, 
    useBytes = TRUE)
  knitr <- gsub(tag.thead, css.thead, knitr, fixed = TRUE, 
    useBytes = TRUE)
  knitr <- gsub(tag.arc, css.arc, knitr, fixed = TRUE, useBytes = TRUE)
  if (remove.spaces) {
    knitr <- sju.rmspc(knitr)
    toWrite <- sju.rmspc(toWrite)
    page.content <- sju.rmspc(page.content)
  }
  structure(class = c("sjTable", "view_df"), list(page.style = page.style, 
    page.content = page.content, output.complete = toWrite, 
    header = NULL, knitr = knitr, file = file, show = !no.output, 
    use.viewer = use.viewer))
}

view_df2(arab4, hide.progress = T)
```
