==> 02_regressions.Rmd <==
---
title: "regressions"
output: html_notebook
---

# Load in Data and Packages

```{r}
pacman::p_load(tidyverse, haven, magrittr, car, psych, sjPlot, sjstats, sjmisc, lme4, binoculaR, MuMIn)

load("data/arab_reg.Rdata")

# arab_reg %<>% 
#   mutate(islamism = range01(islamism)) %>% 
#   mutate(age = ifelse(age == 1, NA, age)) %>% 
#   mutate(age = range01(age))

# table(arab_reg$cntry, arab_reg$governorate)

# arab_reg %>% 
#   filter(cntry == "Egypt") %>% 
#   select(governorate) %>% 
#   unique() %>% 
#   arrange(governorate)

```

# LMER


```{r}
model0 <- arab_reg %>% 
  lme4::lmer(islamism ~ 1 + (1|cntry), data = ., weights = wt)

icc(model0)

model1 <- lmer(islamism ~ female + age + #work + income + educ + 
                 year_2012 + year_2013 + year_2014 + 
                # globalism + personalpiety + patriarchalvalues + liberalislam + 
                 (1|cntry), data = arab_reg, weights = wt)

model2 <- lmer(islamism ~ female + age + work + income + educ + 
                 year_2012 + year_2013 + year_2014 + 
                # globalism + personalpiety + patriarchalvalues + liberalislam + 
                 (1|cntry), data = arab_reg, weights = wt)

model3 <- lmer(islamism ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + #personalpiety + patriarchalvalues + liberalislam + 
                 (1|cntry), data = arab_reg, weights = wt)


model4 <- lmer(islamism ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + #patriarchalvalues + liberalislam + 
                 (1|cntry), data = arab_reg, weights = wt)

model5 <- lmer(islamism ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + patriarchalvalues + #liberalislam + 
                 (1|cntry), data = arab_reg, weights = wt)

model6 <- lmer(islamism ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + patriarchalvalues + liberalislam + 
                 (1|cntry), data = arab_reg, weights = wt)

texreg::screenreg(list(model1, model2, model3, model4, model5, model6))
```



# Interactions

```{r}

# model7 <- lmer(islamism ~ female + work + income + age + educ + 
#                  year_2012 + year_2013 + year_2014 + 
#                  globalism + personalpiety + patriarchalvalues + liberalislam +
#                  personalpiety*work +
#                  (1|cntry), data = arab_reg, weights = wt)

# model8 <- lmer(islamism ~ female + work + income + age + educ + 
#                  year_2012 + year_2013 + year_2014 + 
#                  globalism + personalpiety + patriarchalvalues + liberalislam +
#                  personalpiety*income +
#                  (1|cntry), data = arab_reg, weights = wt)
# 
# model9 <- lmer(islamism ~ female + work + income + age + educ + 
#                  year_2012 + year_2013 + year_2014 + 
#                  globalism + personalpiety + patriarchalvalues + liberalislam +
#                  personalpiety*educ +
#                  (1|cntry), data = arab_reg, weights = wt)
# 
# model10 <- lmer(islamism ~ female + work + income + age + educ + 
#                  year_2012 + year_2013 + year_2014 + 
#                  globalism + personalpiety + patriarchalvalues + liberalislam +
#                  personalpiety*globalism +
#                  (1|cntry), data = arab_reg, weights = wt)


model11 <- lmer(islamism ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + patriarchalvalues + liberalislam +
                 personalpiety*patriarchalvalues +
                 (1|cntry), data = arab_reg, weights = wt)

model12 <- lmer(islamism ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + patriarchalvalues + liberalislam +
                 personalpiety*liberalislam +
                 (1|cntry), data = arab_reg, weights = wt)

#texreg::screenreg(list(model6, model7, model8, model9))

# arab_reg %>% 
#   drop_na(islamism) -> aa

# 18508/21915
# 1-0.8445357

anova(model6, model12)
```



```{r}
texreg::texreg(list(model1, model2, model3, 
                    model4, model5, model6, 
                    model11, model12),
          label = "regressions",
          #caption = "Bolded coefficients, custom notes, three digits.",
          float.pos = "!h", bold = 0.10,
       #   custom.note = "Coefficients with $p < 0.05$ in \\textbf{bold}.",
          digits = 2, leading.zero = FALSE, 
       stars = c(0.001, 0.01, 0.05, 0.1),
          symbol = "\\dagger", #reorder.coef = c(1:6, 8:12, 13:17, 7, 18:22),
          groups = list("\\textbf{Control Variables}" = 2:6,
                        "\\textbf{Variables of Interest}" = 7:13,
                        "\\textbf{Interactions}" = 14:15), 
       reorder.coef = c(1:10, 12, 11, 13:15),
          custom.coef.names =  c("Intercept",
                                 "Sex (Male/Female)",
                                 "Age",
                                 "Year 2012 (0/1)",
                                 "Year 2013 (0/1)",
                                 "Year 2014 (0/1)",
                                 "Employment (0/1)",
                                 "Financial Security",
                                 "Education",
                                 "Parochialism",
                                 "Personal Piety",
                                 "Patriarchal Values",
                                 "Liberal Islam",
                                 "Personal Piety $\\times$ Patriarchal Values",
                                 "Personal Piety $\\times$ Liberal Islam"), custom.note = "Models show unstandardized b-coefficients. All variables normalized (0-1). Reference category for year dummies is 2016. Data weighted to nationally representative samples."
       )
```

```{r}
plot_model(model6, show.values = T, 
           show.p = T, sort.est = T, 
           value.offset = 0.4, 
           vline.color = "gray90", axis.labels = c("Liberal Islam",
                                                   "Age",
                                                   "Education",
                                                   "Year 2014 (0/1)",
                                                   "Employment (0/1)",
                                                   "Year 2012 (0/1)",
                                                   "Financial Security",
                                                   "Sex (Male/Female)",
                                                   "Year 2013 (0/1)", 
                                                   "Parochialism",
                                                   "Personal Piety",
                                                   "Patriarchal Values")) +
  ggthemes::theme_hc() +
  ggtitle("")

ggsave(filename = "text/images/coefplot.png", width = 12, height = 7)

plot_model(model11 ,type = "int", axis.lim = c(.1,0.6), grid.breaks = .1) +
  ggthemes::theme_hc() +
  ggtitle("") +
  ylab("Support for Islamism") +
  xlab("Personal Piety") +
  scale_color_discrete("Patriarchal Values") -> relig_pat
plot_model(model12 ,type = "int", axis.lim = c(.1,0.6)) +
  ggthemes::theme_hc() +
  ggtitle("") +
  ylab("Support for Islamism")  +
  xlab("Personal Piety") +
  scale_color_discrete("Liberal Islam") -> relig_lib

cowplot::plot_grid(relig_pat, relig_lib)

ggsave(filename = "text/images/interactions.png", width = 12, height = 5)



model13 <- lmer(islamism ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + patriarchalvalues + liberalislam +
                 patriarchalvalues*personalpiety +
                 (1|cntry), data = arab_reg, weights = wt)

model14 <- lmer(islamism ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + patriarchalvalues + liberalislam +
                 liberalislam*personalpiety +
                 (1|cntry), data = arab_reg, weights = wt)



plot_model(model13 ,type = "int", axis.lim = c(.1,0.6), grid.breaks = .1) +
  ggthemes::theme_hc() +
  ggtitle("") +
  ylab("Support for Islamism") +
  xlab("Patriarchal Values") +
  scale_color_discrete("Personal Piety") -> relig_pat2
plot_model(model14 ,type = "int", axis.lim = c(.1,0.6)) +
  ggthemes::theme_hc() +
  ggtitle("") +
  ylab("Support for Islamism")  +
  xlab("Liberal Islam") +
  scale_color_discrete("Personal Piety") -> relig_lib2

cowplot::plot_grid(relig_pat2, relig_lib2)

ggsave(filename = "text/images/interactions2.png", width = 12, height = 5)

```

# vifs

```{r}
vif.mer <- function (fit) {
  ## adapted from rms::vif
  
  v <- vcov(fit)
  nam <- names(fixef(fit))
  
  ## exclude intercepts
  ns <- sum(1 * (nam == "Intercept" | nam == "(Intercept)"))
  if (ns > 0) {
    v <- v[-(1:ns), -(1:ns), drop = FALSE]
    nam <- nam[-(1:ns)]
  }
  
  d <- diag(v)^0.5
  v <- diag(solve(v/(d %o% d)))
  names(v) <- nam
  v
}

vif.mer(model12)
```


# GLMER

```{r}
model1a <- glmer(religparty ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + patriarchalvalues + liberalislam +
                 (1|cntry), data = arab_reg, weights = wt,
              family = binomial)

model1b <- glmer(religparty ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + patriarchalvalues + liberalislam +
                 globalism*personalpiety +
                 (1|cntry), data = arab_reg, weights = wt,
              family = binomial)

model1c <- glmer(religparty ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + patriarchalvalues + liberalislam +
                 patriarchalvalues*personalpiety +
                 (1|cntry), data = arab_reg, weights = wt,
              family = binomial)

model1d <- glmer(religparty ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + patriarchalvalues + liberalislam +
                 liberalislam*personalpiety +
                 (1|cntry), data = arab_reg, weights = wt,
              family = binomial)

texreg::screenreg(list(model1a, model1b, model1c, model1d))
```

# brms

```{r}
library(brms)

bm1 <- brms::brm(islamism|weights(wt) ~ female + work + income + age + educ + 
                 year_2012 + year_2013 + year_2014 + 
                 globalism + personalpiety + patriarchalvalues + liberalislam + 
                 (1|cntry), data = arab_reg)

# devtools::install_github("systats/lolR")

save(bm1, file = "data/bm1.Rdata")


load("data/bm1.Rdata")

plot(marginal_effects(bm1))

library(tidybayes)

add_predicted_samples(arab_reg %>% sample_n(20), bm1) -> pred_samples
add_fitted_samples(arab_reg, bm1, n = 100) -> fit_samples

library(ggridges)

pred_samples %>%
  ggplot(aes(x = pred, y = personalpiety)) +
  geom_density_ridges()


library(modelr)

m_mpg <- brms::brm(mpg ~ hp, data = mtcars)

mtcars %>% 
  modelr::data_grid(hp = modelr::seq_range(hp, 101), am) %>% 
  add_predicted_samples(m_mpg)


fit_samples %>%
  ungroup() %>% 
  select(liberalislam, personalpiety, patriarchalvalues, estimate, .iteration) %>% 
  gather("variable", "value", -estimate, -.iteration) %>% 
  # mutate(variable = 
  #          case_when(
  #            variable == "govsat" ~ "Government Satisfaction",
  #            variable == "globalism" ~ "Trust in Global Governance",
  #            variable == "lrscale" ~ "Left-Right Scale"
  #          )) %>% 
  # mutate(category = 
  #          case_when(
  #            category == "1" ~ "Establishment",
  #            category == "2" ~ "Progressive Populism",
  #            category == "3" ~ "Traditionalist Populism"
  #          )) %>% 
  mutate(estimate = round(estimate * 100, 2)) %>% 
 ggplot(aes(x = value, y = estimate)) +
  geom_smooth(aes(y = estimate, group = .iteration), alpha = 0.25, 
              se = F, size = 0.000000000000025, method = "lm") +
  facet_grid(~variable, scales = "free") +
    scale_y_continuous(breaks = seq(0,100, by = 5),
                     labels = paste0(seq(0,100, by = 5),"%")) + 
    ggthemes::theme_hc() +
    labs(y = "Probability", x = "Score") +
  ggthemes::scale_colour_gdocs("Type") + 
    guides(color = F) +
  # cowplot::panel_border() +# and a border around each panel
  theme(panel.spacing = unit(2, "lines"),
        strip.text = element_text(size = 11)) 
    # scale_x_continuous(breaks = seq(0,10, by

ggsave(filename = "slides/images/bayesia_fun2.png", width = 12, height = 9)


pred_samples %>% 
  ggplot(aes(personalpiety)) +
  geom_jitter(aes(personalpiety, pred), alpha = 0.1) + 
  stat_lineribbon(aes(y = pred), .prob = c(.99, .95, .8, .5), alpha = 0.5) +
  scale_fill_brewer() 

```
