==> 04_dataviz.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# data & packages

```{r}
load("data/arab_reg.Rdata")
pacman::p_load(tidyverse)

range01 <- function(x){(x - min(x, na.rm = T)) / (max(x, na.rm = T) - min(x, na.rm = T))}

range14 <- function(x){(range01(x)*3)+1}
```

# overview

```{r}
minus_5 <- function(x) {
  5-x
}

cols <- c('#eff3ff','#bdd7e7','#6baed6','#2171b5')

arab_reg %>% 
  select(islamistparties, islamistgov, religleaders, religleadersinfl) %>% 
  # filter(4 < islamistgov)
  # mutate_all(range14) %>% 
  mutate_all(minus_5) %>% 
  gather() %>% 
  mutate(key = case_when(
    key == "islamistparties" ~ "Only Islamist Parties",
    key == "islamistgov" ~ "Islamist Government (no elections)",
    key == "religleaders" ~ "Religious Leaders hold office",
    key == "religleadersinfl" ~ "Religious Leaders influence decisions")) -> descriptives



descriptives %>% 
  group_by(key, value) %>% 
  tally() %>%
  rename(total = n) %>% 
  na.omit() %>% 
  left_join(descriptives %>% 
              group_by(key) %>% tally) %>% 
  mutate(perc = round((total / n)*100, 2)) %>% 
  ungroup() -> percentrage

percentrage %>% 
  mutate(key = fct_reorder(key, perc, .desc = T)) %>%
  mutate(value = case_when(
    value == 4 ~ "Not suitable at all\\ \n Strongly disagree", 
    value == 3 ~ "Somewhat suitable\\ \n Disagree",  
    value == 2 ~ "Suitable\\ \n Agree",   
    value == 1 ~ "Completely suitable\\ \n Strongly agree"
  )) %>% 
  mutate(value = factor(value, levels = c("Not suitable at all\\ \n Strongly disagree",
                                          "Somewhat suitable\\ \n Disagree", 
                                          "Suitable\\ \n Agree",   
                                          "Completely suitable\\ \n Strongly agree"))) %>% 
  ggplot(aes(y = perc, x = key, fill = as.factor(value))) + 
  geom_bar(position = "fill", stat = "identity", width = 0.4) +
  coord_flip() +
  ggthemes::theme_hc() +
  scale_fill_manual("", values = rev(cols))  +
  geom_text(aes(label = paste0(perc,"%")),
              position = position_fill(vjust=0.5), size = 3.5, color = "black")  + 
    scale_y_continuous(labels = scales::percent_format()) +
  xlab("") +
  ylab("") +
  guides(fill = guide_legend(reverse = T))

ggsave(filename = "images/descriptives.png", width = 13)
```


Country Descriptives


```{r}
arab %>% 
  ggplot(aes(islamism, fill = islamism)) +
  geom_density() +
  facet_wrap(~cntry, scales = "free") +
    ggthemes::theme_hc() 

arab %>% 
  mutate(cntry = fct_reorder(cntry, islamism, .desc = T)) %>%
  ggplot(aes(islamism, cntry)) +
  ggridges::geom_density_ridges()
```

# Map Preprare

```{r}
library(httr)     # getting data
library(rgdal)    # working with shapefile
library(dplyr)    # awesome data manipulation
library(readr)    # faster reading of CSV data
library(stringi)  # string manipulation
library(stringr)  # string manipulation
library(tidyr)    # reshaping data
library(grid)     # for 'unit'
library(scales)   # for 'percent'
library(ggplot2)  # plotting
library(ggthemes) # theme_map
 
# this ensures you only download the shapefile once and hides
# errors and warnings. remove `try` and `invisible` to see messages
try(invisible(GET("http://www.pewglobal.org/wp-content/lib/js/world-geo.json",
                  write_disk("world-geo.json"))), silent=TRUE)
 
# use ogrListLayers("world-geo.json") to see file type & 
# layer info to use in the call to readOGR
#ogrListLayers("world-geo.json")
world <- rgdal::readOGR("world-geo.json")
world_wt <- spTransform(world, CRS("+proj=robin"))
world_map <- fortify(world_wt)

world_map %>%
  left_join(data_frame(id=rownames(world@data), name=world@data$name)) %>%
  select(-id) %>%
  rename(id=name) -> world_map

world_map %<>% 
  mutate(id_new = countrycode::countrycode(id, "country.name", "country.name")) %>% 
  mutate(id = ifelse(is.na(id_new), id, id_new)) %>% 
  select(-id_new)

save(world_map, file = "data/world_map.Rdata")

```


# real mapping

```{r}
library(sjmisc)
load("data/world_map.Rdata")

world_map %>% 
  filter(long > -1711303.61 & long < 5420923) %>% 
  filter(lat > 440000 & lat < 30469918) %>% 
  mutate(region = countrycode::countrycode(id, "country.name", "region")) %>% 
  filter(str_detect(region, "Asia") | str_detect(region, "Africa")) %>% 
  filter(id %nin% c("Iran", "Turkey")) %>% 
  select(-region) -> mena_map

arab_reg %>% 
  mutate(islamism = range01(islamism)) %>% 
  group_by(cntry) %>% 
  summarise(islamism = mean(islamism, na.rm=T)) %>% 
  #mutate(region = countrycode::countrycode(cntry, "country.name", "region")) %>% 
  mutate(id = countrycode::countrycode(cntry, "country.name", "country.name")) %>% 
  left_join(mena_map) -> arab_map





cname <- aggregate(cbind(long, lat) ~ cntry, data = arab_map, FUN=function(x)mean(range(x))) %>% 
  rbind(data.frame(
    cntry = "Palestine",
    long = 3210742.9,
    lat = 3522516
  )) %>% 
  left_join(arab_map %>% 
  group_by(cntry) %>% 
  summarise(islamism = mean(range(islamism, na.rm=T)))) %>% 
  mutate(islamism= round(islamism, 2)) 
  

mena_map %>% 
  ggplot() +
  geom_map(map = mena_map,
         aes(x = long, y = lat, group = group, map_id = id),
         color = "#7f7f7f", fill = "gray80", size = 0.15) +
  geom_map(data = arab_map, 
           map = mena_map,
        aes(map_id  = id, 
            fill = islamism), size = 0.01) + 
  theme_map() +
  coord_equal() +
  viridis::scale_fill_viridis("Average Islamism Score", 
                              option = "D", discrete = F, end = .8) + 
  theme(
    legend.justification = c(1, 0),
    legend.position = c(0.49, 0.14),
    legend.title = element_text(size = 14),
    axis.ticks.length = unit(3, "cm"),
    legend.direction = "horizontal",
    legend.background = element_rect(fill = "transparent")) +
  guides(fill = guide_colorbar(barwidth = 13, barheight = 0.5,
                title.position = "top", title.hjust = 0.5,
                label.theme = element_text(colour = "black", size = 5, angle = 0))) +
  ggrepel::geom_label_repel(data = cname, aes(long, lat, label = paste0(cntry, ":", islamism)))



ggsave(filename = "images/map_islamism.png", height = 12, width = 9) 

# mena_map %>% select(region) %>% table()
```


# descriptives

```{r}

arab_reg %>% 
  drop_na(educ) %>% 
 # mutate(educ = as.factor(educ)) %>% 
  group_by(educ) %>% 
  summarise(sd_islamism = sd(islamism, na.rm = T) %>% 
            round(., 2),
            islamism = mean(islamism, na.rm = T) %>% 
            round(., 2)) -> educ_means
  
arab_reg %>%
  drop_na(islamism) %>% 
  ggplot(aes(educ, islamism, group = as.factor(educ), fill = as.factor(educ))) +
  geom_violin(alpha = 0.7)  + 
#  scale_fill_manual(values = cols) +
  geom_boxplot(width=0.06, fill = "white")  + 
 # geom_hline(yintercept = mean(mlogit_ess$lrscale, na.rm = T), linetype = 2) + #line at base mean
  ggthemes::theme_hc() +
  scale_fill_grey("", start = 0.2, end = 0.6)  +
  # ggthemes::scale_fill_fivethirtyeight() + 
  ggpubr::stat_compare_means(label = "p.format") + # Add pairwise comparisons p-value
  ggpubr::stat_compare_means(label.y = 9.3, label.x = 3.4) +    # Add global p-value
  guides(fill=F) +
  xlab("Education") +
  ylab("Support for Islamism") +
  # ggtitle("Left-Right Scale by Party Support") +
  #labs(caption = "Source: ESS Data Round 5 - 8; N = 87238") +
  coord_flip() +
  geom_text(data = educ_means,
            aes(label = islamism), nudge_y = 0.06) +
  theme(axis.text=element_text(size = 10), 
        title = element_text(size = 18, face = "bold"), 
        plot.caption = element_text(size = 10),
        axis.title=element_text(size = 10),
        axis.text.y = element_text(face = "bold")) -> education_plot

ggsave(filename = "images/education.png", width = 13, height = 9)
 


arab_reg %>% 
  drop_na(age) %>% 
  mutate(age_cat = sjmisc::split_var(age, n = 5)) %>% 
 # mutate(age = as.factor(age)) %>% 
  group_by(age_cat) %>% 
  summarise(sd_islamism = sd(islamism, na.rm = T) %>% 
            round(., 2),
            islamism = mean(islamism, na.rm = T) %>% 
            round(., 2)) -> age_means
  
arab_reg %>%
  drop_na(age) %>% 
  mutate(age_cat = sjmisc::split_var(age, n = 5)) %>% 
  drop_na(islamism) %>% 
  ggplot(aes(age_cat, islamism, group = as.factor(age_cat), fill = as.factor(age_cat))) +
  geom_violin(alpha = 0.7)  + 
#  scale_fill_manual(values = cols) +
  geom_boxplot(width=0.2, fill = "white")  + 
 # geom_hline(yintercept = mean(mlogit_ess$lrscale, na.rm = T), linetype = 2) + #line at base mean
  ggthemes::theme_hc() +
  scale_fill_grey("", start = 0.2, end = 0.6)  +
  # ggthemes::scale_fill_fivethirtyeight() + 
  # ggpubr::stat_compare_means(label = "p.format") + # Add pairwise comparisons p-value
  # ggpubr::stat_compare_means(label.y = 9.3, label.x = 3.4) +    # Add global p-value
  guides(fill=F) +
  xlab("Age") +
  ylab("Support for Islamism") +
  # ggtitle("Left-Right Scale by Party Support") +
  #labs(caption = "Source: ESS Data Round 5 - 8; N = 87238") +
  coord_flip() +
  geom_text(data = age_means,
            aes(label = islamism), nudge_y = 0.06) +
  theme(axis.text=element_text(size = 10), 
        title = element_text(size = 18, face = "bold"), 
        plot.caption = element_text(size = 10),
        axis.title=element_text(size = 10),
        axis.text.y = element_text(face = "bold")) -> age_plot

ggsave(filename = "images/ageation.png", width = 13, height = 9)
 


arab_reg %>% 
  drop_na(female) %>% 
 # mutate(age = as.factor(age)) %>% 
  group_by(female) %>% 
  summarise(sd_islamism = sd(islamism, na.rm = T) %>% 
            round(., 2),
            islamism = mean(islamism, na.rm = T) %>% 
            round(., 2)) -> female_means
  
arab_reg %>%
  drop_na(female) %>% 
  drop_na(islamism) %>% 
  ggplot(aes(female, islamism, group = as.factor(female), fill = as.factor(female))) +
  geom_violin(alpha = 0.7)  + 
#  scale_fill_manual(values = cols) +
  geom_boxplot(width=0.2, fill = "white")  + 
 # geom_hline(yintercept = mean(mlogit_ess$lrscale, na.rm = T), linetype = 2) + #line at base mean
  ggthemes::theme_hc() +
  scale_fill_grey("", start = 0.2, end = 0.6)  +
  # ggthemes::scale_fill_fivethirtyeight() + 
  # ggpubr::stat_compare_means(label = "p.format") + # Add pairwise comparisons p-value
  # ggpubr::stat_compare_means(label.y = 9.3, label.x = 3.4) +    # Add global p-value
  guides(fill=F) +
  xlab("Sex (Male / Female)") +
  ylab("Support for Islamism") +
  # ggtitle("Left-Right Scale by Party Support") +
  #labs(caption = "Source: ESS Data Round 5 - 8; N = 87238") +
  coord_flip() +
  geom_text(data = female_means,
            aes(label = islamism), nudge_y = 0.06) +
  theme(axis.text=element_text(size = 10), 
        title = element_text(size = 18, face = "bold"), 
        plot.caption = element_text(size = 10),
        axis.title=element_text(size = 10),
        axis.text.y = element_text(face = "bold")) -> sex_plot

ggsave(filename = "images/femaleation.png", width = 13, height = 9)




arab_reg %>% 
  drop_na(work) %>% 
 # mutate(age = as.factor(age)) %>% 
  group_by(work) %>% 
  summarise(sd_islamism = sd(islamism, na.rm = T) %>% 
            round(., 2),
            islamism = mean(islamism, na.rm = T) %>% 
            round(., 2)) -> work_means
  
arab_reg %>%
  drop_na(work) %>% 
  drop_na(islamism) %>% 
  ggplot(aes(work, islamism, group = as.factor(work), fill = as.factor(work))) +
  geom_violin(alpha = 0.7)  + 
#  scale_fill_manual(values = cols) +
  geom_boxplot(width=0.2, fill = "white")  + 
 # geom_hline(yintercept = mean(mlogit_ess$lrscale, na.rm = T), linetype = 2) + #line at base mean
  ggthemes::theme_hc() +
  scale_fill_grey("", start = 0.2, end = 0.6)  +
  # ggthemes::scale_fill_fivethirtyeight() + 
  # ggpubr::stat_compare_means(label = "p.format") + # Add pairwise comparisons p-value
  # ggpubr::stat_compare_means(label.y = 9.3, label.x = 3.4) +    # Add global p-value
  guides(fill=F) +
  xlab("Employed (0/1)") +
  ylab("Support for Islamism") +
  # ggtitle("Left-Right Scale by Party Support") +
  #labs(caption = "Source: ESS Data Round 5 - 8; N = 87238") +
  coord_flip() +
  geom_text(data = work_means,
            aes(label = islamism), nudge_y = 0.06) +
  theme(axis.text=element_text(size = 10), 
        title = element_text(size = 18, face = "bold"), 
        plot.caption = element_text(size = 10),
        axis.title=element_text(size = 10),
        axis.text.y = element_text(face = "bold")) -> work_plot

ggsave(filename = "images/workation.png", width = 13, height = 9)


plot2by2 <- cowplot::plot_grid(education_plot, age_plot, sex_plot, work_plot,
                      labels = c("A", "B", "C", "D"), ncol = 2)

cowplot::save_plot("text/images/plot2by2.png", plot2by2,
          ncol = 2, # we're saving a grid plot of 2 columns
          nrow = 2, # and 2 rows
          base_height = 6,
          base_width = 10,
          # each individual subplot should have an aspect ratio of 1.3
          base_aspect_ratio = 1.3
          )
 
```

```{r}

arab_reg %>% 
  drop_na(income) %>% 
 # mutate(age = as.factor(age)) %>% 
  group_by(income) %>% 
  summarise(sd_islamism = sd(islamism, na.rm = T) %>% 
            round(., 2),
            islamism = mean(islamism, na.rm = T) %>% 
            round(., 2)) -> income_means
  
arab_reg %>%
  drop_na(income) %>% 
  drop_na(islamism) %>% 
  ggplot(aes(income, islamism, group = as.factor(income), fill = as.factor(income))) +
  geom_violin(alpha = 0.7)  + 
#  scale_fill_manual(values = cols) +
  geom_boxplot(width=0.2, fill = "white")  + 
 # geom_hline(yintercept = mean(mlogit_ess$lrscale, na.rm = T), linetype = 2) + #line at base mean
  ggthemes::theme_hc() +
  scale_fill_grey("", start = 0.2, end = 0.6)  +
  # ggthemes::scale_fill_fivethirtyeight() + 
  # ggpubr::stat_compare_means(label = "p.format") + # Add pairwise comparisons p-value
  # ggpubr::stat_compare_means(label.y = 9.3, label.x = 3.4) +    # Add global p-value
  guides(fill=F) +
  xlab("ageation") +
  ylab("Support for Islamism") +
  # ggtitle("Left-Right Scale by Party Support") +
  #labs(caption = "Source: ESS Data Round 5 - 8; N = 87238") +
  coord_flip() +
  geom_text(data = income_means,
            aes(label = islamism), nudge_y = 0.3) +
  theme(axis.text=element_text(size = 10), 
        title = element_text(size = 18, face = "bold"), 
        plot.caption = element_text(size = 10),
        axis.title=element_text(size = 10),
        axis.text.y = element_text(face = "bold"))

ggsave(filename = "images/incomeation.png", width = 13, height = 9)
 
```


```{r}
arab_reg %>% 
  ggplot(aes(personalpiety, islamism, color = as.factor(liberalislam))) +
 # geom_jitter() +
  geom_smooth(method = "lm", se = F) +
  ggpubr::stat_cor()
```

# summary table

```{r}
arab_reg %>% 
  #select(-islamistparties:-religleadersinfl, -cntry) %>% 
  select(`Islamism` = islamism,
         `Sex (Male/Female)` = female, 
         `Age` = age, 
         `Year 2012 (0/1)` = year_2012, 
         `Year 2013 (0/1)` = year_2013, 
         `Year 2014 (0/1)` = year_2014, 
         `Financial Security` = income,
         `Employment (0/1)` = work, 
         `Education` = educ, 
         `Parochialism` = globalism, 
         `Personal Piety` = personalpiety, 
         `Patriarchal Values` = patriarchalvalues,
         `Liberal Islam` = liberalislam) %>% 
  descr() %>% 
  as.data.frame() %>% 
  select(variable, n, mean, sd, md, min, max) %>% 
  mutate(max = round(max)) %>% 
  mutate(n = round(n)) %>% 
  mutate_if(is.numeric, ~round(.,2)) %>% 
  mutate_if(is.numeric, as.character) %>% 
  rename(Variable = variable, 
         N = n,
         Mean = mean,
         SD = sd,
         Median = md,
         Min = min,
         Max = max) %>% 
  xtable::xtable() %>% 
  print(., booktabs = T)

hist(arab_reg$liberalislam)
``` 

# heatmap

```{r}
ggheatmap <- function(.data) {
  
 library(reshape2)
 
 cormat <- round(cor(.data, use = "pairwise.complete.obs"),3)
 
 # Get upper triangle of the correlation matrix
 get_upper_tri <- function(cormat){
     cormat[lower.tri(cormat)] <- NA
     return(cormat)
   }
 
 reorder_cormat <- function(cormat){
 # Use correlation between variables as distance
 dd <- as.dist((1-cormat)/2)
 hc <- hclust(dd)
 cormat <- cormat[hc$order, hc$order]
 }
 
 # Reorder the correlation matrix
 #cormat <- reorder_cormat(cormat)
 upper_tri <- get_upper_tri(cormat)
 # Melt the correlation matrix
 melted_cormat <- melt(upper_tri, na.rm = TRUE) %>% 
   mutate(value = sprintf('%.2f', value, 2)) %>% 
   mutate(value = as.numeric(value))
 # Create a ggheatmap
 ggplot(melted_cormat, aes(Var2, Var1, fill = value)) +
  geom_tile(color = "white")+
  scale_fill_gradient2(low = "blue", high = "red", mid = "white", 
    midpoint = 0, limit = c(-1,1), space = "Lab", 
     name="Pearson Correlation\n") +
  ggthemes::theme_hc()+ # minimal theme
  theme(axis.text.x = element_text(angle = 45, vjust = 1, 
     size = 12, hjust = 1))+
 # coord_fixed()  + 
 geom_text(aes(Var2, Var1, label = value), color = "black", size = 4) +
 theme(
   axis.title.x = element_blank(),
   axis.title.y = element_blank(),
   panel.grid.major = element_blank(),
   panel.border = element_blank(),
   panel.background = element_blank(),
   axis.ticks = element_blank(),
   legend.justification = c(1, 0),
   legend.position = c(0.6, 0.7),
   legend.title = element_text(size = 20),
  # axis.ticks.length = unit(2, "cm"),
   legend.direction = "horizontal")+
   guides(fill = guide_colorbar(barwidth = 15, barheight = 1,
                 title.position = "top", title.hjust = 0.5))
}

arab_reg %>% 
    select(`Islamism` = islamism,
         `Sex (Male/Female)` = female, 
         `Age` = age, 
         `Year 2012 (0/1)` = year_2012, 
         `Year 2013 (0/1)` = year_2013, 
         `Year 2014 (0/1)` = year_2014, 
         `Financial Security` = income,
         `Employment (0/1)` = work, 
         `Education` = educ, 
         `Parochialism` = globalism, 
         `Personal Piety` = personalpiety, 
         `Patriarchal Values` = patriarchalvalues,
         `Liberal Islam` = liberalislam)  %>% 
  ggheatmap()

ggsave(filename = "images/heatmap.png", width = 10, height = 10)
```
